[vpsAdminOS] Remount / as shared, recursively
It seems to be the standard to have filesystems mounted as shared. For
some reason, vpsAdminOS has its mounts private. This is not an issue on
the vpsAdminOS host, but it is a problem for containers. systemd expects
the filesystems to be shared for its features to work, namely
propagation of host mounts to running services or the LoadCredential=
feature.